clip_phogpt/requirements.txt
einops 
nltk==3.9.1
rouge-score
pycocoevalcap
pillow
transformers 
datasets 
torch 
torchvision
